feat(anta.cli): Implement run-template debug command
