+++ anta/cli/debug/utils.py
@@ -5,7 +5,7 @@
 Utils functions to use with anta.cli.cli module.
 """
 
-from anta.models import AntaTest, AntaTestCommand, AntaTestTemplate
+from anta.models import AntaTest, AntaTestCommand
 
 
 class RunArbitraryCommand(AntaTest):
@@ -18,26 +18,6 @@ class RunArbitraryCommand(AntaTest):
     name = "Run aributrary EOS command"
     description = "To be used only with anta debug commands"
     commands = [AntaTestCommand(command="show version")]
-    categories = ["debug"]
-
-    @AntaTest.anta_test
-    def test(self) -> None:
-        """
-        Fake test function
-        CLI should only call self.collect()
-        """
-
-
-class RunArbitraryTemplateCommand(AntaTest):
-
-    """
-    Run an EOS command and return result
-    Based on AntaTest to build relevant output for pytest
-    """
-
-    name = "Run aributrary EOS command"
-    description = "To be used only with anta debug commands"
-    template = AntaTestTemplate(template="show interfaces {ifd}")
     categories = ["debug"]
 
     @AntaTest.anta_test
